Move index_folder to image_folder. Tweak AppState to be mutable, using the mutexy pattern in Tauri's documentation to do so. Tweak page to wire up an event to listen to app state changes.
NOTE THAT
My current plan is to implement the folder-search in a kinda "parallel track" to the rest of the code, rather than try and do the hard-cognitive work of abstracting out the data layer. Once I have the parallel track, I reckon it'll be easier for me practically to then unify the two tracks.
(This is largely based on the energy/productivity I have available)